write the per-tile-hash in graphtilebuilder
