+++ valhalla/mjolnir/util.h
@@ -11,7 +11,6 @@
 
 #include <array>
 #include <atomic>
-#include <filesystem>
 #include <map>
 #include <mutex>
 #include <span>
@@ -350,13 +349,6 @@ bool build_tile_set(const boost::property_tree::ptree& config,
                     const std::vector<std::string>& input_files,
                     const BuildStage start_stage = BuildStage::kInitialize,
                     const BuildStage end_stage = BuildStage::kValidate);
-
-/**
- * Recompute a single tile's 48-bit data hash and store it in the low bits of the header checksum,
- * preserving the tileset build id already in the high bits.
- * @param tile_path filesystem path to the .gph tile
- */
-void set_tile_checksum(const std::filesystem::path& tile_path);
 
 /**
  * Recompute the tileset-wide build id from the per-tile data hashes already stored in each tile
